Add an intermediate `DowncastSend` that `DowncastSync` extends.
`DowncastSend` and `DowncastSync` provide methods to upcast to `Box<Trait + Send>` and `Box<Trait + Send + Sync>`, respectively.

The addition of a new `DowncastSend` trait that `DowncastSync` depends on is probably backward-compatible since no-one else should be able timplement `DowncastSync` due to object-safety rules. But we need to switch to a new major version anyway. So taking advantage of that here to not worry about backward-compatibility edge cases.

Fixes #14.